src/requirements.txt
asgiref==3.6.0
Django==3.2.18
django-pivot==1.9.0
djangorestframework==3.14.0
mysqlclient==2.1.1
pytz==2022.7.1
sqlparse==0.4.3
